Ustalenie częściowe layoutów dla widoku głównego i fragmentów dla ViewPager2
Zmieniono ViewModel dla w. w. fragmentów

Poprawki, refaktoring

+++ app/src/main/java/com/example/towardsgoalsapp/goals/GoalData.kt
@@ -4,5 +4,5 @@ data class GoalData(
     val goalId: Int,
     val goalName: String,
     val goalDescription: String,
-    val progress: Float
+    val progress: Double
     )
